refactor: migrate plan API to custom fetch hooks
