+++ src/entities/plan/api/index.ts
@@ -1,7 +1,3 @@
 export * from './endpoint'
 export * from './queryKey'
-export * from './get-plan'
-export * from './get-plans'
-export * from './post-plan'
-export * from './update-plan'
-export * from './delete-plan'
+export * from './use-plans'
