--- src/entities/plan/api/update-plan.ts
@@ -1,27 +1,0 @@
-import { customAxios } from '@/src/shared/api'
-import { PlanPayloadType } from '../model'
-import { useMutation, useQueryClient } from '@tanstack/react-query'
-import { PLAN_QUERY_KEY } from './queryKey'
-import { PLAN_URL } from './endpoint'
-
-export const updatePlan = async (id: string, payload: PlanPayloadType) => {
-  try {
-    const response = await customAxios.patch(PLAN_URL.detail(id), payload)
-    return response.data.results
-  } catch (error) {
-    console.error(error)
-    throw error
-  }
-}
-
-export const useUpdatePlan = (plan_id: string) => {
-  const queryClient = useQueryClient()
-
-  return useMutation({
-    mutationFn: (data: PlanPayloadType) => updatePlan(plan_id, data),
-    onSuccess: () => {
-      queryClient.refetchQueries({ queryKey: PLAN_QUERY_KEY.detail(plan_id) })
-      queryClient.invalidateQueries({ queryKey: PLAN_QUERY_KEY.all })
-    },
-  })
-}
